engine boilerplate and vertices rendering
