+++ Quickie/Quickie.sln
@@ -1,9 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2013
-VisualStudioVersion = 12.0.40629.0
+VisualStudioVersion = 12.0.21005.1
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Quickie", "Quickie.vcxproj", "{96D6BC99-F15F-4514-ACB2-74D8FCE2BB06}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "PlanetCollision", "PlanetCollision.vcxproj", "{0259B800-2046-467E-94A2-D13B6F5210BC}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{A47F2171-4D9C-4CDE-AD89-1A030C4A5FF9}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +13,10 @@ Global
 		Release|Win32 = Release|Win32
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{96D6BC99-F15F-4514-ACB2-74D8FCE2BB06}.Debug|Win32.ActiveCfg = Debug|Win32
-		{96D6BC99-F15F-4514-ACB2-74D8FCE2BB06}.Debug|Win32.Build.0 = Debug|Win32
-		{96D6BC99-F15F-4514-ACB2-74D8FCE2BB06}.Release|Win32.ActiveCfg = Release|Win32
-		{96D6BC99-F15F-4514-ACB2-74D8FCE2BB06}.Release|Win32.Build.0 = Release|Win32
+		{0259B800-2046-467E-94A2-D13B6F5210BC}.Debug|Win32.ActiveCfg = Debug|Win32
+		{0259B800-2046-467E-94A2-D13B6F5210BC}.Debug|Win32.Build.0 = Debug|Win32
+		{0259B800-2046-467E-94A2-D13B6F5210BC}.Release|Win32.ActiveCfg = Release|Win32
+		{0259B800-2046-467E-94A2-D13B6F5210BC}.Release|Win32.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
